# swagger_docs.yaml
openapi: 3.0.0
info:
  title: SSStudio API
  version: 1.0.0
  description: API for anonymization, synthesis, and balancing of datasets.

servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        201:
          description: User registered successfully
        400:
          description: Missing username or password
        409:
          description: Username already exists

  /login:
    post:
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        200:
          description: Login successful
        401:
          description: Invalid credentials

  /ping:
    get:
      summary: Health check endpoint
      responses:
        200:
          description: Server is up

  /test:
    get:
      summary: Test authentication
      security:
        - bearerAuth: []
      responses:
        200:
          description: JWT is valid

  /anonymize:
    post:
      summary: Anonymize data
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
                configs:
                  type: object
                  properties:
                    model:
                      type: string
              required: [data, configs]
      responses:
        200:
          description: Successfully anonymized data
        400:
          description: Missing or incorrect parameters
        500:
          description: Internal server error

  /synthesize:
    post:
      summary: Synthesize data using model
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
                configs:
                  type: object
                  properties:
                    model:
                      type: string
                    num_rows:
                      type: integer
                    categorical_cols:
                      type: array
                      items:
                        type: string
                    epochs:
                      type: integer
              required: [data, configs]
      responses:
        200:
          description: Successfully synthesized data
        400:
          description: Missing or incorrect parameters
        500:
          description: Internal server error

  /balance:
    post:
      summary: Balance dataset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: array
                  items:
                    type: object
                configs:
                  type: object
                  properties:
                    model:
                      type: string
                    target_column:
                      type: string
              required: [data, configs]
      responses:
        200:
          description: Successfully balanced data
        400:
          description: Missing or incorrect parameters
        500:
          description: Internal server error